Added .editorconfig for easier editing.
Editor Config allows local per-project configuration to be expressed in a single
file, which makes it easy for one to work across many repos with many projects,
each of which may have its own preferred indentation, spacing, EOL, encoding,
etc. styles.

This change adds a 2-space indent for Ruby and YAML files; other rules can be
added quite easily. The file has a pointer to the project's homepage which
explains how to install editor-specific plugins to handle this config and how
to extend this config further.