Introduce Concourse pipeline
This pipeline performs two functions.

First, it performs an integration test of the
upgrade-google-client branch whenever new changes appear there.

Second, it runs integration tests against every PR with the
`integrate` label. This allows whitelisted PRs to be tested
automatically before being merged. Status is updated on the
PR when the test completes.

When #210 lands, switch the pipeline to work against master.