Explicitly set rubocop TargetRubyVersion to 2.1
This is the lowest version of ruby travis is run against, so hound
should match to prevent version mismatches.

This will prevent hound from requesting changes which are
incompatible with versions we support. An example of such a
feature is safe navigation, not available before 2.3. ie,
github.com/fog/fog-google/pull/231#pullrequestreview-49858308