Upgrade Storage models to >= 0.11
Models have been minimally changed to work against rewritten
requests. Model tests have been added.

This follows up the previous changes to Storage requests
bf540a0

Semantics of public_url have been changed. We no longer
validate the ACL as it is impossible to mutate
those ACLs as a result of the below changes. Hence,
we simply return the url which would provide access to
the resource if it was public.

Requests and their tests have been modify to validate
creating objects and buckets with predefined ACLs.

The ACL mutations, which operated mainly using predefined
ACLs, have been removed as they were previously unimplemented.
When I attempted to implement them, I ran into a semantics
mismatch with what we are allowed to use predefined ACLs for.
You can create an object or bucket with a predefined ACL.
You cannot update an object or bucket with a predefined ACL.
I confirmed this behavior is acknowledged internally
and will not be changed.

This was mentioned in google-cloud-ruby
GoogleCloudPlatform/google-cloud-ruby#381
(Final paragraph)

To reproduce this ACL issue:
Create a new bucket in GCP console, add an object using upload.
Try to use the 'Objects: patch' method with the API Explorer
to change to a predefined ACL. You will encounter the following error

```
"code": 409,
  "message": "Cannot provide both a predefinedAcl and access controls."
```

https://cloud.google.com/storage/docs/json_api/v1/objects/patch