Avoid duplicate GET requests when retrieving body
Previously the response body would be loaded twice in memory whenever
`GoogleJSON::File#body` was called. This was happening because inside
the ternary condition, `file.body` was recursively calling itself twice
even though the body was already loaded. We rewrite this method to be
simpler and to have a proper base condition.

Closes https://github.com/fog/fog-google/issues/541